VC6 support (targets NT 3.51)

+++ gfx/video_driver.c
@@ -2321,7 +2321,7 @@ void video_driver_frame(const void *data, unsigned width,
          snprintf(frames_text,
                sizeof(frames_text),
                STRING_REP_UINT64,
-               (unsigned long long)video_driver_frame_count);
+               (uint64_t)video_driver_frame_count);
 
          strlcat(video_driver_window_title,
                frames_text,
@@ -2337,7 +2337,7 @@ void video_driver_frame(const void *data, unsigned width,
                "FPS: %6.1f || %s: " STRING_REP_UINT64,
                last_fps,
                msg_hash_to_str(MSG_FRAMES),
-               (unsigned long long)video_driver_frame_count);
+               (uint64_t)video_driver_frame_count);
    }
    else
    {
