+++ menu/cbs/menu_cbs_get_value.c
@@ -525,9 +525,9 @@ static void menu_action_setting_disp_set_label_perf_counters_common(
 #else
          "%llu ticks, %llu runs.",
 #endif
-         ((unsigned long long)counters[offset]->total /
-          (unsigned long long)counters[offset]->call_cnt),
-         (unsigned long long)counters[offset]->call_cnt);
+         ((uint64_t)counters[offset]->total /
+          (uint64_t)counters[offset]->call_cnt),
+         (uint64_t)counters[offset]->call_cnt);
 }
 
 static void general_disp_set_label_perf_counters(
